Executing a mapping should execute all now.
So it iterates backwards, creating partially applied functions that hold
their count and previous partial function. So each mapping target can
execute it's dependency when required.

So if you call 5d10w, it will execute d and pass it the number 5, PLUS
the w motion with the prefilled count argument. So d can just call w and
it will know how many it should delete.

I don't like to use profanities in my commits but: Holy shit, this is
pants on head crazy. I love it.